porting 2022 oct 05 1732

+++ src/win_dock_style.rs
@@ -1,7 +1,9 @@
 #![allow(non_snake_case)]
 
+use crate::window_dock_style_color::ImGuiWindowDockStyleCol_COUNT;
+
 #[derive(Default,Debug,Clone)]
 pub struct ImGuiWindowDockStyle
 {
-    pub Colors: [u32;ImGuiWindowDockStyleCol_COUNT]
+    pub Colors: [u32; ImGuiWindowDockStyleCol_COUNT as usize]
 }
